+++ requirements.txt
@@ -1,0 +1,3 @@
+Faker==25.2.0
+Flask==3.0.3
+python_bcrypt==4.1.3
