+++ trabalho integrador/client/templates/buscar.html
@@ -13,6 +13,8 @@
         <label for="id"></label><input type="number" id="id" name="id"  placeholder="Insira ID"><br>
         <label for="nome"></label><input type="text" id="nome" name="nome" placeholder="Insira nome"><br>
         <label for="cpf"></label><input type="number" id="cpf" name="cpf"  placeholder="Insira CPF"><br>
+        <br>
+        <label for="senha"></label><input type="password" id="senha" name="senha" placeholder="Confirme sua senha"><br>
 
         <br>
         <input type="submit" value="Buscar">
