+++ trabalho integrador/client/templates/resultado.html
@@ -11,10 +11,10 @@
     <table border="1">
         <tr>
             <th>ID</th>
-            <th>Nome/Usuario</th>
-            <th>CPF/Senha</th>
-            <th>Histórico/Permissão</th>
-            <th>Score/Nível</th>
+            <th>Nome</th>
+            <th>CPF</th>
+            <th>Histórico</th>
+            <th>Score</th>
         </tr>
         {% for registro in resultado %}
             <tr>
