--- trabalho integrador/templates/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Index</title>
-</head>
-<body>
-    <h1>Bem-vindo</h1>
-    <p>Por favor, faça login para continuar.</p>
-    <a href="/login">Login</a>
-</body>
-</html>
